add gmp-lib to requires_add
Autoconf will indirectly depend on perl and perl-Math-BigInt-GMP.
In cases of containerized environments like docker and when the host
system is running a "generic" x86-64 system ld will try to load libgmp
from /usr/lib64 and not /usr/lib64/haswell, in that case we need to
have the generic one.

This isssue was observed while building a clear linux package inside
a container.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>